requirements.txt
wheel
setuptools
requests
PyYAML
loguru
# python-dotenv
